+++ src/Oro/Bundle/PayPalBundle/Method/Config/AbstractPayPalConfig.php
@@ -6,21 +6,9 @@ use Oro\Bundle\PaymentBundle\Method\Config\ParameterBag\AbstractParameterBagPaym
 
 abstract class AbstractPayPalConfig extends AbstractParameterBagPaymentConfig implements PayPalConfigInterface
 {
-    const LABEL_KEY = 'label';
-    const SHORT_LABEL_KEY = 'short_label';
-    const ADMIN_LABEL_KEY = 'admin_label';
-    const PAYMENT_METHOD_IDENTIFIER_KEY = 'payment_method_identifier';
     const CREDENTIALS_KEY  = 'credentials';
     const PURCHASE_ACTION_KEY  = 'purchase_action';
     const TEST_MODE_KEY  = 'test_mode';
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __construct(array $parameters)
-    {
-        parent::__construct($parameters);
-    }
 
     /**
      * {@inheritdoc}