+++ src/Oro/Bundle/ProductBundle/Entity/Product.php
@@ -1301,6 +1301,7 @@ class Product extends ExtendProduct implements
         return [
             'id' => $this->getId(),
             'product_units' => $this->getAvailableUnitsPrecision(),
+            'unit' => $this->getPrimaryUnitPrecision()->getProductUnitCode(),
             'name' => $this->getDefaultName() ? $this->getDefaultName()->getString() : '',
         ];
     }