+++ src/Oro/Bundle/ProductBundle/Tests/Behat/Context/FeatureContext.php
@@ -36,6 +36,7 @@ use Oro\Bundle\WarehouseBundle\SystemConfig\WarehouseConfig;
  * @SuppressWarnings(PHPMD.TooManyPublicMethods)
  * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
  * @SuppressWarnings(PHPMD.ExcessivePublicCount)
+ * @SuppressWarnings(PHPMD.ExcessiveClassLength)
  */
 class FeatureContext extends OroFeatureContext implements OroPageObjectAware, KernelAwareContext
 {
@@ -707,6 +708,92 @@ class FeatureContext extends OroFeatureContext implements OroPageObjectAware, Ke
     }
 
     /**
+     * Assert that embedded block contains specified product with specified element.
+     * Example: Then should see "Low Inventory" for "PSKU1" product in the "New Arrivals Block"
+     *
+     * @Then /^(?:|I )should see "(?P<elementName>[^"]*)" for "(?P<SKU>[^"]*)" product in the "(?P<blockName>[^"]+)"$/
+     */
+    public function iShouldSeeElementForTheFollowingProductsInEmbeddedBlock($elementName, $SKU, $blockName)
+    {
+        $block = $this->createElement($blockName);
+        self::assertTrue($block->isValid(), sprintf('Embedded block "%s" was not found', $blockName));
+
+        $productItem = $this->findElementContains('ProductItem', $SKU, $block);
+        self::assertNotNull($productItem, sprintf('product with SKU "%s" not found', $SKU));
+
+        if ($this->isElementVisible($elementName, $productItem)) {
+            return;
+        }
+
+        self::assertNotFalse(
+            stripos($productItem->getText(), $elementName),
+            sprintf('text or element "%s" for product with SKU "%s" is not present or not visible', $elementName, $SKU)
+        );
+    }
+
+    /**
+     * Assert that embedded block does not contain specified product with specified element.
+     * Example: Then should not see "Low Inventory" for "PSKU1" product in the "New Arrivals Block"
+     *
+     * @Then /^(?:|I )should not see "(?P<element>[^"]*)" for "(?P<SKU>[^"]*)" product in the "(?P<blockName>[^"]+)"$/
+     */
+    public function iShouldNotSeeElementForTheFollowingProductsInEmbeddedBlock($element, $SKU, $blockName)
+    {
+        $block = $this->createElement($blockName);
+        self::assertTrue($block->isValid(), sprintf('Embedded block "%s" was not found', $blockName));
+
+        $productItem = $this->findElementContains('ProductItem', $SKU, $block);
+        self::assertNotNull($productItem, sprintf('product with SKU "%s" not found', $SKU));
+
+        $textAndElementPresentedOnPage = $this->isElementVisible($element, $productItem)
+            || stripos($productItem->getText(), $element);
+
+        self::assertFalse(
+            $textAndElementPresentedOnPage,
+            sprintf('text or element "%s" for product with SKU "%s" is present or visible', $element, $SKU)
+        );
+    }
+
+    /**
+     * @Then /^(?:|I )should see "(?P<elementName>[^"]*)" for "(?P<SKU>[^"]*)" line item$/
+     */
+    public function shouldSeeForLineItem($elementName, $SKU)
+    {
+        $productItem = $this->findElementContains('ProductLineItem', $SKU);
+        self::assertNotNull($productItem, sprintf('line item with SKU "%s" not found', $SKU));
+
+        if ($this->isElementVisible($elementName, $productItem)) {
+            return;
+        }
+
+        self::assertNotFalse(
+            stripos($productItem->getText(), $elementName),
+            sprintf(
+                'text or element "%s" for line item with SKU "%s" is not present or not visible',
+                $elementName,
+                $SKU
+            )
+        );
+    }
+
+    /**
+     * @Then /^(?:|I )should not see "(?P<elementName>[^"]*)" for "(?P<SKU>[^"]*)" line item$/
+     */
+    public function shouldNotSeeForLineItem($elementName, $SKU)
+    {
+        $productItem = $this->findElementContains('ProductLineItem', $SKU);
+        self::assertNotNull($productItem, sprintf('line item with SKU "%s" not found', $SKU));
+
+        $textAndElementPresentedOnPage = $this->isElementVisible($elementName, $productItem)
+            || stripos($productItem->getText(), $elementName);
+
+        self::assertFalse(
+            $textAndElementPresentedOnPage,
+            sprintf('text or element "%s" for line item with SKU "%s" is present or visible', $elementName, $SKU)
+        );
+    }
+
+    /**
      * @Then /^(?:|I )click "([^"]*)" for "([^"]*)" product$/
      */
     public function clickElementForSelectedProduct($elementName, $SKU)
@@ -729,7 +816,7 @@ class FeatureContext extends OroFeatureContext implements OroPageObjectAware, Ke
     }
 
     /**
-     * @Given /^I should see "([^"]*)" product$/
+     * @Given /^(?:|I )should see "([^"]*)" product$/
      */
     public function iShouldSeeInSearchResults($productSku)
     {
@@ -738,7 +825,7 @@ class FeatureContext extends OroFeatureContext implements OroPageObjectAware, Ke
     }
 
     /**
-     * @Then /^I should not see "([^"]*)" product$/
+     * @Then /^(?:|I )should not see "([^"]*)" product$/
      *
      * @param string $productSku
      */
@@ -988,19 +1075,20 @@ class FeatureContext extends OroFeatureContext implements OroPageObjectAware, Ke
     }
 
     /**
-     * @param string $elementName
-     * @param NodeElement $productItem
-     * @return bool
+     * Select a value for product attribute on product update form
+     * Example: I fill in product attribute "Color" with "Red"
+     *
+     * @When /^(?:|I )fill in product attribute "(?P<field>(?:[^"]|\\")*)" with "(?P<value>(?:[^"]|\\")*)"$/
      */
-    protected function isElementVisible($elementName, $productItem)
+    public function fillProductAttribute($field, $value)
     {
-        if ($this->hasElement($elementName)) {
-            $element = $this->createElement($elementName, $productItem);
-            if ($element->isValid() && $element->isVisible()) {
-                return true;
-            }
-        }
+        $field = $this->fixStepArgument($field);
+        $value = $this->fixStepArgument($value);
+        $form = $this->createElement('OroForm');
+        $value = $form->normalizeValue($value);
 
-        return false;
+        $form
+            ->find('css', sprintf('[name="oro_product[%s]"]', $field))
+            ->setValue($value);
     }
 }