+++ src/Oro/Bundle/ShoppingListBundle/Resources/public/js/app/views/product-add-to-shopping-list-view.js
@@ -185,6 +185,7 @@ define(function(require) {
 
             if (hasLineItems) {
                 var $removeButton =  $(this.options.removeButtonTemplate(shoppingList));
+                $removeButton.find('a').attr('data-intention', 'remove');
                 buttons.push($removeButton);
             }
         },
@@ -240,10 +241,16 @@ define(function(require) {
             return hasUnits ? foundShoppingList : null;
         },
 
-        _onQuantityEnter: function(e) {
-            var ENTER_KEY_CODE = 13;
+        findShoppingListLineItemByUnit: function(shoppingList, unit) {
+            var foundShoppingList = this.findShoppingListById(shoppingList);
+            if (!foundShoppingList) {
+                return null;
+            }
+            return _.find(foundShoppingList.line_items, {unit: unit});
+        },
 
-            if (e.keyCode === ENTER_KEY_CODE && this.dropdownWidget.main.data('intention') === 'update') {
+        _onQuantityEnter: function(e) {
+            if (e.keyCode === 13) {
                 if (!this.dropdownWidget.validateForm()) {
                     return;
                 }
@@ -251,14 +258,32 @@ define(function(require) {
                 this.model.set({
                     quantity: parseInt($(e.target).val(), 10)
                 });
-                this._saveLineItem();
+
+                var $button = this.findMainButton();
+                var url = $button.data('url');
+                var intention = $button.data('intention');
+                var currentShoppingList = this.findCurrentShoppingList();
+                var formData = this.$form.serialize();
+                var urlOptions = {
+                    shoppingListId: currentShoppingList.id,
+                    productId: this.model.get('id')
+                };
+
+                if (this.model.has('parentProduct')) {
+                    urlOptions.parentProductId = this.model.get('parentProduct');
+                }
+
+                this._saveToShoppingList(url, urlOptions, formData, intention);
             }
         },
 
         findCurrentShoppingList: function() {
-            return this.shoppingListCollection.find(function(model) {
-                return model.get('is_current') === true;
-            }).toJSON() || null;
+            return this.shoppingListCollection.find({is_current: true}).toJSON() || null;
+        },
+
+        findShoppingList: function(id) {
+            id = parseInt(id, 10);
+            return this.shoppingListCollection.find({id: id}).toJSON() || null;
         },
 
         onClick: function(e) {
@@ -269,12 +294,14 @@ define(function(require) {
             var url = $button.data('url');
             var intention = $button.data('intention');
             var formData = this.$form.serialize();
-            var urlOptions = {};
 
             if (!this.dropdownWidget.validateForm()) {
                 return;
             }
 
+            var urlOptions = {
+                shoppingListId: $button.data('shoppinglist').id
+            };
             if (this.model) {
                 urlOptions.productId = this.model.get('id');
                 if (this.model.has('parentProduct')) {
@@ -282,15 +309,7 @@ define(function(require) {
                 }
             }
 
-            if (intention === 'new') {
-                this._createNewShoppingList(url, urlOptions, formData);
-            } else if (intention === 'update') {
-                this._saveLineItem();
-            } else {
-                var shoppingList = $button.data('shoppinglist');
-                urlOptions.shoppingListId = shoppingList.id;
-                this._addProductToShoppingList(url, urlOptions, formData);
-            }
+            this._saveToShoppingList(url, urlOptions, formData, intention);
         },
 
         updateLabel: function($button, shoppingList, hasLineItems) {
@@ -318,6 +337,34 @@ define(function(require) {
                 .attr('data-intention', intention);
 
             return $button;
+        },
+
+        /**
+         * Save quantity product to shopping list
+         * Create new shopping list
+         * Update/Add to shopping list
+         * Remove from shopping list
+         *
+         * @param {String} url
+         * @param {Object} urlOptions
+         * @param {Object} formData
+         * @param {String} action
+         * @private
+         */
+        _saveToShoppingList: function(url, urlOptions, formData, action) {
+            switch (action) {
+                case 'new':
+                    this._createNewShoppingList(url, urlOptions, formData);
+                    break;
+                case 'update':
+                    this._saveLineItem(url, urlOptions, formData);
+                    break;
+                case 'remove':
+                    this._removeLineItem(url, urlOptions, formData);
+                    break;
+                default:
+                    this._addLineItem(url, urlOptions, formData);
+            }
         },
 
         _setEditLineItem: function(lineItemId, setFirstLineItem) {
@@ -358,7 +405,7 @@ define(function(require) {
             var dialog = new ShoppingListCreateWidget({});
             dialog.on('formSave', _.bind(function(response) {
                 urlOptions.shoppingListId = response.savedId;
-                this._addProductToShoppingList(url, urlOptions, formData);
+                this._addLineItem(url, urlOptions, formData);
             }, this));
             dialog.render();
         },
@@ -375,26 +422,9 @@ define(function(require) {
                 url: routing.generate(url, urlOptions),
                 data: formData,
                 success: function(response) {
-                    mediator.execute('hideLoading');
-                    if (response && response.message) {
-                        mediator.execute(
-                            'showFlashMessage', (response.hasOwnProperty('successful') ? 'success' : 'error'),
-                            response.message
-                        );
-
-                        self.model.set('shopping_lists', response.product.shopping_lists, {silent: true});
-                        self.model.trigger('change:shopping_lists');
-                        if (!self.shoppingListCollection.find({id: response.shoppingList.id})) {
-                            self.shoppingListCollection.add(_.defaults(response.shoppingList, {is_current: true}), {
-                                silent: true
-                            });
-                        }
-                        self.shoppingListCollection.trigger('change', {
-                            refresh: true
-                        });
-                    }
+                    mediator.trigger('shopping-list:line-items:update-response', self.model, response);
                 },
-                error: function(xhr) {
+                complete: function() {
                     mediator.execute('hideLoading');
                 }
             });
@@ -412,14 +442,25 @@ define(function(require) {
             this.model.trigger('focus:quantity');
         },
 
-        _saveLineItem: function(urlOptions, formData) {
+        _addLineItem: function(url, urlOptions, formData) {
+            this._addProductToShoppingList(url, urlOptions, formData);
+        },
+
+        _removeLineItem: function(url, urlOptions, formData) {
+            this._addProductToShoppingList(url, urlOptions, formData);
+        },
+
+        _saveLineItem: function(url, urlOptions, formData) {
             if (this.model && !this.model.get('line_item_form_enable')) {
                 return;
             }
             mediator.execute('showLoading');
 
+            var shoppingList = this.findShoppingList(urlOptions.shoppingListId);
+            var lineItem = this.findShoppingListLineItemByUnit(shoppingList, this.model.get('unit'));
+
             var savePromise = this.saveApiAccessor.send({
-                id: this.editLineItem.id
+                id: lineItem.id
             }, {
                 quantity: this.model.get('quantity'),
                 unit: this.model.get('unit')
@@ -427,8 +468,8 @@ define(function(require) {
 
             savePromise
                 .done(_.bind(function(response) {
-                    this.editLineItem.quantity = response.quantity;
-                    this.editLineItem.unit = response.unit;
+                    lineItem.quantity = response.quantity;
+                    lineItem.unit = response.unit;
                     this.shoppingListCollection.trigger('change', {
                         refresh: true
                     });