+++ CHANGELOG.md
@@ -3,14 +3,22 @@
 ### Added
 #### CheckoutBundle
 * Added `CheckoutLineItem` and `CheckoutSubtotal` entities. They will be used in `Checkout` entity to store data. Previously for these purposes used line items and subtotals of Checkout source entity (`ShoppingList` or `QuoteDemand` entities).
-
+#### OrderBundle
+* Added Previously purchased products functionality. [Documentation](./src/Oro/Bundle/OrderBundle/Resources/doc/previously-purchased-products.md)
+#### RFPBundle
+* Added new email template `request_create_confirmation`. It will be send when guest customer user create new request for quote.
+* Added new twig function `rfp_products` that returns list of request products (formatted) for current request for quote. Can be used in email templates.
 #### WebsiteSearchBundle
 * Added interface `Oro\Bundle\WebsiteSearchBundle\Attribute\Type\SearchableAttributeTypeInterface` that should be implemented in case new type of arguments added.
+
 ### Changed
 #### CheckoutBundle
 * Entity `Oro\Bundle\CheckoutBundle\Entity\Checkout`:
     * no longer implements `Oro\Bundle\PricingBundle\SubtotalProcessor\Model\LineItemsNotPricedAwareInterface`;
     * implements `Oro\Bundle\CurrencyBundle\Entity\CurrencyAwareInterface`.
+
+#### InventoryBundle
+* Added Low Inventory Highlights functionality.[Documentation](./src/Oro/Bundle/InventoryBundle/Resources/doc/low_inventory_highlights.md)
 
 #### ProductBundle
 * Updated website search configuration file `Oro/Bundle/ProductBundle/Resources/config/oro/website_search.yml`: